Move refresh and selection methods to list VM
